Move Wheelchair mapping to StopAndStationMapper
	- Simply move business logic and take Wheelchair enum as input on StopMapper
